practice-omi/api_practice/pyproject.toml
[tool.poetry]
name = "api_practice"
version = "0.1.0"
description = "django rest framework practice"
authors = ["Farhan Ishraq <omifarhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
djangorestframework = "^3.12.4"
Pygments = "^2.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
